Updated createdUser list after user update and delete

+++ tests/acceptance/features/bootstrap/Provisioning.php
@@ -1240,6 +1240,7 @@ trait Provisioning {
 			$this->getAdminPassword(),
 			$this->ocsApiVersion
 		);
+		$this->createdUsers[$user]["email"] = $email;
 	}
 
 	/**
@@ -1310,6 +1311,7 @@ trait Provisioning {
 			$targetUser,
 			$email
 		);
+		$this->createdUsers[$targetUser]["email"] = $email;
 	}
 
 	/**
@@ -1330,6 +1332,7 @@ trait Provisioning {
 			$email
 		);
 		$this->theHTTPStatusCodeShouldBeSuccess();
+		$this->createdUsers[$targetUser]["email"] = $email;
 	}
 
 	/**
@@ -1353,6 +1356,7 @@ trait Provisioning {
 		$this->adminChangesTheDisplayNameOfUserUsingKey(
 			$user, 'displayname', $displayname
 		);
+		$this->createdUsers[$user]["displayname"] = $displayname;
 	}
 
 	/**
@@ -1384,6 +1388,7 @@ trait Provisioning {
 		);
 		$this->setResponse($response);
 		$this->theDisplayNameReturnedByTheApiShouldBe($displayname);
+		$this->createdUsers[$user]["displayname"] = $displayname;
 	}
 
 	/**
@@ -1407,6 +1412,7 @@ trait Provisioning {
 		$this->adminChangesTheDisplayNameOfUserUsingKey(
 			$user, 'display', $displayname
 		);
+		$this->createdUsers[$user]["displayname"] = $displayname;
 	}
 
 	/**
@@ -1460,6 +1466,7 @@ trait Provisioning {
 		$this->userChangesTheDisplayNameOfUserUsingKey(
 			$requestingUser, $targetUser, 'displayname', $displayName
 		);
+		$this->createdUsers[$targetUser]["displayname"] = $displayName;
 	}
 
 	/**
@@ -1483,6 +1490,7 @@ trait Provisioning {
 		$this->userChangesTheDisplayNameOfUserUsingKey(
 			$requestingUser, $targetUser, 'display', $displayName
 		);
+		$this->createdUsers[$targetUser]["displayname"] = $displayName;
 	}
 
 	/**
@@ -1501,6 +1509,7 @@ trait Provisioning {
 			$requestingUser, $targetUser, 'displayname', $displayName
 		);
 		$this->theHTTPStatusCodeShouldBeSuccess();
+		$this->createdUsers[$targetUser]["displayname"] = $displayName;
 	}
 	/**
 	 *
@@ -2013,6 +2022,22 @@ trait Provisioning {
 			$this->createdUsers[$user] = $userData;
 		} elseif ($this->currentServer === 'REMOTE') {
 			$this->createdRemoteUsers[$user] = $userData;
+		}
+	}
+
+	/**
+	 * @param string $user
+	 * @param string $key key for userData used in createdList
+	 * @param string $value value for update
+	 *
+	 * @return void
+	 */
+	public function updateUserInCreatedUsersList($user, $key, $value) {
+		$user = $this->normalizeUsername($user);
+		if ($this->currentServer === 'LOCAL') {
+			$this->createdUsers[$user][$key] = $value;
+		} elseif ($this->currentServer === 'REMOTE') {
+			$this->createdRemoteUsers[$user][$key] = $value;
 		}
 	}
 
@@ -3937,11 +3962,15 @@ trait Provisioning {
 		$previousServer = $this->currentServer;
 		$this->usingServer('LOCAL');
 		foreach ($this->createdUsers as $user => $userData) {
-			$this->deleteUser($user);
+			if ($userData["shouldExist"]) {
+				$this->deleteUser($user);
+			}
 		}
 		$this->usingServer('REMOTE');
 		foreach ($this->createdRemoteUsers as $remoteUser => $userData) {
-			$this->deleteUser($remoteUser);
+			if ($userData["shouldExist"]) {
+				$this->deleteUser($remoteUser);
+			}
 		}
 		$this->usingServer($previousServer);
 	}
@@ -3956,11 +3985,15 @@ trait Provisioning {
 		$previousServer = $this->currentServer;
 		$this->usingServer('LOCAL');
 		foreach ($this->createdGroups as $group => $groupData) {
-			$this->cleanupGroup($group);
+			if ($groupData["shouldExist"] and $groupData["possibleToDelete"]) {
+				$this->cleanupGroup($group);
+			}
 		}
 		$this->usingServer('REMOTE');
 		foreach ($this->createdRemoteGroups as $remoteGroup => $groupData) {
-			$this->cleanupGroup($remoteGroup);
+			if ($groupData["shouldExist"] and $groupData["possibleToDelete"]) {
+				$this->cleanupGroup($remoteGroup);
+			}
 		}
 		$this->usingServer($previousServer);
 	}
